refactor: create parent class for event 'checkout_onepage_controller_success_action'
